+++ src/main/scala/fs2/aws/sqs/SqsConfig.scala
@@ -1,0 +1,5 @@
+package fs2.aws.sqs
+
+import eu.timepit.refined.types.string.TrimmedString
+
+case class SqsConfig(queueName: TrimmedString)
